# evcar-news/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "AdminRoles_list",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete list (not paginated): meta is page 1 of 1 with pageSize = total.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RoleDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Roles with their permissions and user counts",
        "tags": [
          "admin: roles"
        ]
      }
    },
    "/api/v1/admin/roles/{key}": {
      "get": {
        "operationId": "AdminRoles_get",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RoleDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "One role",
        "tags": [
          "admin: roles"
        ]
      }
    },
    "/api/v1/admin/roles/{key}/permissions": {
      "put": {
        "description": "403 OWNER_ONLY for non-owners; 409 ROLE_NOT_EDITABLE for the owner role; 422 UNKNOWN_PERMISSION.",
        "operationId": "AdminRoles_setPermissions",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RoleDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope\n\nError envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Replace the permissions of a role (owners only)",
        "tags": [
          "admin: roles"
        ]
      }
    },
    "/api/v1/admin/permissions": {
      "get": {
        "operationId": "AdminPermissions_list",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete list (not paginated): meta is page 1 of 1 with pageSize = total.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PermissionDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Every permission key with its group and description",
        "tags": [
          "admin: roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "EV Car News API",
    "description": "REST API of EV Car News (evcar.news). Conventions:\n- Prefix /api/v1; admin routes under /api/v1/admin (permission-guarded), personal routes under /api/v1/me.\n- Language: ?lang=ar|en overrides Accept-Language (default ar). Market: ?market=EG overrides X-Market.\n- Single resource: {\"data\": {...}}; lists: {\"data\": [...], \"meta\": {page, pageSize, total, totalPages}}.\n- Errors: {\"error\": {\"code\", \"message\" (localized), \"details\"?, \"requestId\"}}; validation → 422 VALIDATION_FAILED.\n- Money: {\"amount\": \"decimal string\", \"currency\": \"EGP\"}; missing values are null (never 0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "example": "بعض الحقول غير صالحة. راجع التفاصيل."
          },
          "details": {
            "type": "object",
            "description": "Code-specific details (e.g. field errors for VALIDATION_FAILED)."
          },
          "requestId": {
            "type": "string",
            "example": "0b8f6c1e-5a1c-4c0e-9d6b-2f3b9b4d1a77"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string",
            "example": "editor"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean"
          },
          "permissions": {
            "description": "Effective permission keys (the owner role always has all of them).",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissionsEditable": {
            "type": "boolean",
            "description": "False for the owner role (always every permission)."
          },
          "userCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "key",
          "nameAr",
          "nameEn",
          "description",
          "isSystem",
          "permissions",
          "permissionsEditable",
          "userCount"
        ]
      },
      "SetRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "example": [
              "articles.read",
              "articles.create"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "PermissionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "articles.publish"
          },
          "group": {
            "type": "string",
            "example": "articles"
          },
          "descriptionAr": {
            "type": "string",
            "nullable": true
          },
          "descriptionEn": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "key",
          "group",
          "descriptionAr",
          "descriptionEn"
        ]
      },
...
    }
  }
}